% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\alias{confusionTable}
\title{Calculate Classification Accuracy}
\usage{
accuracy(actual, predicted, rank)

confusionTable(actual, predicted, rank)
}
\arguments{
\item{actual}{data.frame with the actual classification hierarchy.}

\item{predicted}{data.frame with the predicted classification hierarchy.}

\item{rank}{rank at which the accuracy should be evaluated.}
}
\value{
The accuracy or a confusion table.
}
\description{
Calculate the classification accuracy at a given phylogenetic level.
}
\examples{

\dontrun{
# the download may take a while
if (!file.exists("data")) {
  if (!file.exists("data.tgz"))
    download.file(paste0("https://master.dl.sourceforge.net/project/",
      "rdp-classifier/rdp-classifier/data.tgz"), "data.tgz", mode = "wb")
  untar("data.tgz")
}

classifier <- rdp("data/classifier/16srrna")

seq <- readRNAStringSet(system.file("examples/RNA_example.fasta",
    package = "rRDP"
))

### decode the actual classification
actual <- decode_Greengenes(names(seq))

### use RDP to predict the classification
pred <- predict(classifier, seq)

### calculate accuracy
confusionTable(actual, pred, "genus")
accuracy(actual, pred, "genus")
}
}
\keyword{model}
