% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_rRDP-package.R
\docType{package}
\name{rRDP-package}
\alias{rRDP}
\alias{rRDP-package}
\title{rRDP: Interface to the RDP Classifier}
\description{
This package installs and interfaces the naive Bayesian classifier for 16S rRNA sequences developed by the Ribosomal Database Project (RDP). With this package the classifier trained with the standard training set can be used or a custom classifier can be trained.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhahsler/rRDP/}
  \item Report bugs at \url{https://github.com/mhahsler/rRDP/issues}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID})

Authors:
\itemize{
  \item Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID})
  \item Nagar Anurag
}

}
\keyword{internal}
